Update target for the makefile

+++ Makefile
@@ -60,7 +60,7 @@ else
 	printf = /bin/echo -e "$(1)$(3)$(subst $(OBJ_DIR)/,,$(2))$(NORMAL)"
 endif
 
-.PHONY: all debug release nocolor clean clean_deps
+.PHONY: all debug release nocolor clean clean_deps update
 
 all: $(BINARY)
 debug: $(BINARY)
@@ -73,18 +73,23 @@ clean:
 clean_deps:
 	@find $(OBJ_DIR) -name '*.d*' | xargs rm -f
 
+update:
+	@git pull && git submodule update
+
 ifneq ($(MAKECMDGOALS),clean) 						 # don't create .d files just to remove them...
 ifneq ($(MAKECMDGOALS),clean_deps)
-$(shell mkdir -p $(OBJ_DIR))							 # make OBJ_DIR for a helper file
-$(shell echo 0 > $(OBJ_DIR)/objs_current)  # reset the built object counter
-$(shell touch .MakefileObjsTotal) 				 # force the helper makefile to always run
-$(shell $(MAKE) -f .MakefileObjsTotal all) # count the number of objects to be built
-$(shell rm -rf $(OBJ_DIR)/deps)						 # remove the dir used to sign the first .d file
-$(shell rm -rf .lockdir)									 # reset the lock for .make_get_progress.sh
+ifneq ($(MAKECMDGOALS),update)
+	$(shell mkdir -p $(OBJ_DIR))							 # make OBJ_DIR for a helper file
+	$(shell echo 0 > $(OBJ_DIR)/objs_current)  # reset the built object counter
+	$(shell touch .MakefileObjsTotal) 				 # force the helper makefile to always run
+	$(shell $(MAKE) -f .MakefileObjsTotal all) # count the number of objects to be built
+	$(shell rm -rf $(OBJ_DIR)/deps)						 # remove the dir used to sign the first .d file
+	$(shell rm -rf .lockdir)									 # reset the lock for .make_get_progress.sh
 
-# include the dependency files
--include $(DEPS)
--include $(PRECOMPILED_HEADER_DEP)
+	# include the dependency files
+	-include $(DEPS)
+	-include $(PRECOMPILED_HEADER_DEP)
+endif
 endif
 endif
